# CAB432-Strobe-Server-Py-1.0.0/insomnia/strobe-openapi.yaml
openapi: 3.0.3
info:
  title: Strobe Local API
  version: 1.0.0
  contact:
    name: CAB432 Teaching Team
    email: cab432@qut.edu.au
  description: |
    Local OpenAPI specification for the Strobe Python server.

    This spec is designed for direct import into Insomnia or other API clients and reflects
    the request and response shapes implemented by the FastAPI server in python-server/src.

    Authentication uses email + password.
    Username is system-managed and mirrors the email value.

    Authentication uses JWT bearer tokens. Register or log in first, then send:

    `Authorization: Bearer <token>`
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT token only. Do not include the "Bearer " prefix.
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000
...
    PostIdParam:
      in: path
      name: postId
      required: true
      description: Post identifier.
      schema:
        type: string
        example: post_12345678
    CommentIdParam:
      in: path
      name: commentId
      required: true
      description: Comment identifier.
      schema:
        type: string
        example: comment_12345678
...
    QueryLimit:
      in: query
      name: limit
      required: false
      description: Maximum number of records to return. The server caps this at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    QueryOffset:
      in: query
      name: offset
      required: false
      description: Number of records to skip before returning results.
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Title is required
...
    UserStats:
      type: object
      required: [followers, following, posts]
      properties:
        followers:
          type: integer
          example: 12
        following:
          type: integer
          example: 7
        posts:
          type: integer
          example: 4
    User:
      type: object
      required: [id, username, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake.bradford@example.com
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        role:
          type: string
          enum: [user, moderator]
          example: user
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
    EnrichedUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [stats, currentUserFollows]
          properties:
            stats:
              $ref: "#/components/schemas/UserStats"
            currentUserFollows:
              type: boolean
              example: false
    PostAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake.bradford@example.com
    PostStats:
      type: object
      required: [likes, comments]
      properties:
        likes:
          type: integer
          example: 3
        comments:
          type: integer
          example: 1
    Post:
      type: object
      required:
        [
          id,
          userId,
          title,
          description,
          images,
          status,
          createdAt,
          updatedAt,
          author,
          stats,
          currentUserLiked,
        ]
      properties:
        id:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        status:
          type: string
          enum: [active, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        author:
          $ref: "#/components/schemas/PostAuthor"
        stats:
          $ref: "#/components/schemas/PostStats"
        currentUserLiked:
          type: boolean
          example: false
    CommentAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake.bradford@example.com
    Comment:
      type: object
      required: [id, postId, userId, text, createdAt, updatedAt, author]
      properties:
        id:
          type: string
          example: 770e8400-e29b-41d4-a716-446655440002
        postId:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        text:
          type: string
          example: Great light in that first shot.
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:16:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:16:00.000Z
        author:
          $ref: "#/components/schemas/CommentAuthor"
...
    Moment:
      type: object
      required:
        [id, userId, imageUrl, caption, status, createdAt, expiresAt, updatedAt]
      properties:
        id:
          type: string
          example: moment_1783417870915
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        imageUrl:
          type: string
          example: /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        caption:
          type: string
          example: Afternoon lab break.
        status:
          type: string
          enum: [active, archived, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:30:00.000Z
        expiresAt:
          type: string
          format: date-time
          example: 2026-07-09T09:30:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:30:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        archivedAt:
          type: string
          format: date-time
          nullable: true
          example: null
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        password:
          type: string
          minLength: 6
          example: password123
        role:
          type: string
          enum: [user, moderator]
          default: user
          example: user
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        password:
          type: string
          example: password123
    CreatePostRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
...
    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          example: Great light in that first shot.
    CreateMomentRequest:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          example: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
        caption:
          type: string
          example: Afternoon lab break.
...
    AuthResponse:
      type: object
      required: [message, user, token]
      properties:
        message:
          type: string
          example: Created successfully
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    UserResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
          example: Success
        user:
          $ref: "#/components/schemas/EnrichedUser"
    UserListResponse:
      type: object
      required: [message, users]
      properties:
        message:
          type: string
          example: Success
        users:
          type: array
          items:
            $ref: "#/components/schemas/EnrichedUser"
    PostResponse:
      type: object
      required: [message, post]
      properties:
        message:
          type: string
          example: Success
        post:
          $ref: "#/components/schemas/Post"
    PostListResponse:
      type: object
      required: [message, posts]
      properties:
        message:
          type: string
          example: Success
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
    CommentResponse:
      type: object
      required: [message, comment]
      properties:
        message:
          type: string
          example: Created successfully
        comment:
          $ref: "#/components/schemas/Comment"
    CommentListResponse:
      type: object
      required: [message, comments]
      properties:
        message:
          type: string
          example: Success
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    MomentResponse:
      type: object
      required: [message, moment]
      properties:
        message:
          type: string
          example: Created successfully
        moment:
          $ref: "#/components/schemas/Moment"
    MomentListResponse:
      type: object
      required: [message, moments]
      properties:
        message:
          type: string
          example: Success
        moments:
          type: array
          items:
            $ref: "#/components/schemas/Moment"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Deleted successfully
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ValidationError
            message: Title is required
    Unauthorised:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorised
            message: Invalid or expired token
    Forbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Forbidden
            message: You can only modify your own posts
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NotFound
            message: Post not found
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register user
      description: Creates a new user account and immediately returns a JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: jake.bradford@example.com
              password: password123
              role: user
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                message: User registered successfully
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake.bradford@example.com
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email is already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Conflict
                message: Email already exists
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Login user
      description: Authenticates an existing user and returns a fresh JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: jake.bradford@example.com
              password: password123
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                message: Login successful
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake.bradford@example.com
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Unauthorised
                message: Invalid email or password
  /v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: |
        Returns public user profiles. If a bearer token is supplied, each user also includes
        currentUserFollows based on the authenticated caller.
      parameters:
        - in: query
          name: q
          required: false
          description: Optional case-insensitive username search fragment.
          schema:
            type: string
            example: alex
        - $ref: "#/components/parameters/QueryLimit"
      responses:
        "200":
          description: Matching users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
              example:
                message: Success
                users:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake.bradford@example.com
                    email: jake.bradford@example.com
                    role: user
                    createdAt: 2026-07-08T09:00:00.000Z
                    updatedAt: 2026-07-08T09:00:00.000Z
                    stats:
                      followers: 12
                      following: 7
                      posts: 4
                    currentUserFollows: false
  /v1/users/{id}:
    get:
      tags: [Users]
      operationId: getUserProfile
      summary: Get user profile
      description: Returns one public user profile with follower, following, and post counts.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              example:
                message: Success
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake.bradford@example.com
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                  stats:
                    followers: 12
                    following: 7
                    posts: 4
                  currentUserFollows: false
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Users]
      operationId: deleteOwnAccount
      summary: Delete own account
      description: Deletes the authenticated user's own account and cascades related posts, comments, likes, follows, moments, and uploads.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Account deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/posts:
    post:
      tags: [Posts]
      operationId: createPost
      summary: Create post
      description: Creates a new post for the authenticated user. Description and images are optional.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            example:
              title: Sunrise ride
              description: Cold start, good coffee, and an empty river path.
              images:
                - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
      responses:
        "201":
          description: Post created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              example:
                message: Created successfully
                post:
                  id: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  title: Sunrise ride
                  description: Cold start, good coffee, and an empty river path.
                  images:
                    - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
                  status: active
                  createdAt: 2026-07-08T09:10:00.000Z
                  updatedAt: 2026-07-08T09:10:00.000Z
                  hiddenBy: null
                  hiddenAt: null
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: jake.bradford@example.com
                  stats:
                    likes: 0
                    comments: 0
                  currentUserLiked: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
...
  /v1/posts/{postId}/comments:
    get:
      tags: [Comments]
      operationId: getPostComments
      summary: Get comments for post
      description: Returns comments in ascending creation order.
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      responses:
        "200":
          description: Comments returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
              example:
                message: Success
                comments:
                  - id: 770e8400-e29b-41d4-a716-446655440002
                    postId: 660e8400-e29b-41d4-a716-446655440001
                    userId: 550e8400-e29b-41d4-a716-446655440010
                    text: Great light in that first shot.
                    createdAt: 2026-07-08T09:16:00.000Z
                    updatedAt: 2026-07-08T09:16:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440010
                      username: matthew.mckague@example.com
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Comments]
      operationId: createComment
      summary: Create comment
      description: Adds a comment to a post as the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
            example:
              text: Great light in that first shot.
      responses:
        "201":
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
              example:
                message: Created successfully
                comment:
                  id: 770e8400-e29b-41d4-a716-446655440002
                  postId: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440010
                  text: Great light in that first shot.
                  createdAt: 2026-07-08T09:16:00.000Z
                  updatedAt: 2026-07-08T09:16:00.000Z
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440010
                    username: matthew.mckague@example.com
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/posts/{postId}/comments/{commentId}:
    delete:
      tags: [Comments]
      operationId: deleteComment
      summary: Delete comment
      description: Deletes a comment. Only the comment author can delete it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
      responses:
        "200":
          description: Comment deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Non-author attempted delete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You can only delete your own comments
        "404":
          description: Comment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: Comment not found
  /v1/feed:
    get:
      tags: [Feed]
      operationId: getPersonalFeed
      summary: Get personal feed
      description: Returns posts from accounts the authenticated user follows, newest first.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Feed returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
              example:
                message: Success
                posts:
                  - id: 660e8400-e29b-41d4-a716-446655440011
                    userId: 550e8400-e29b-41d4-a716-446655440010
                    title: Campus coffee stop
                    description: Fast break before the next lab.
                    images: []
                    status: active
                    createdAt: 2026-07-08T09:12:00.000Z
                    updatedAt: 2026-07-08T09:12:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440010
                      username: matthew.mckague@example.com
                    stats:
                      likes: 1
                      comments: 0
                    currentUserLiked: false
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
  /v1/moments:
    post:
      tags: [Moments]
      operationId: createMoment
      summary: Create moment
      description: Creates a new 24-hour moment for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMomentRequest"
            example:
              imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
              caption: Afternoon lab break.
      responses:
        "201":
          description: Moment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentResponse"
              example:
                message: Created successfully
                moment:
                  id: moment_1783417870915
                  userId: 550e8400-e29b-41d4-a716-446655440000
                  imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
                  caption: Afternoon lab break.
                  status: active
                  createdAt: 2026-07-08T09:30:00.000Z
                  expiresAt: 2026-07-09T09:30:00.000Z
                  updatedAt: 2026-07-08T09:30:00.000Z
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
  /v1/moments/feed:
    get:
      tags: [Moments]
      operationId: getMomentFeed
      summary: Get moment feed
      description: Returns active moments from the authenticated user and accounts they follow.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active moments returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomentListResponse"
              example:
                message: Success
                moments:
                  - id: moment_1783417870915
                    userId: 550e8400-e29b-41d4-a716-446655440000
                    imageUrl: /uploads/550e8400-e29b-41d4-a716-446655440000/post_12345678/5f9439f1.jpg
                    caption: Afternoon lab break.
                    status: active
                    createdAt: 2026-07-08T09:30:00.000Z
                    expiresAt: 2026-07-09T09:30:00.000Z
                    updatedAt: 2026-07-08T09:30:00.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
...